requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
open_clip_torch>=2.24.0
easyocr>=1.7.1
opencv-python-headless>=4.9.0
Pillow>=10.3.0
streamlit>=1.33.0
pyttsx3>=2.90
numpy>=1.26.0
accelerate>=0.30.0
google-generativeai>=0.7.0